Changed double to float, Point3d to Point3f, few memory initialization fixes

+++ src/controlUI/Multiple-Plane-JLinkage/additionalSteps.hpp
@@ -37,76 +37,76 @@
  * @param [out] numberOfPlanes - Number of planes we're considering after removing unnecessary planes
  */
 void removeUnnecessaryPlanes(
-		const vector<Point3d> &oldData,
+		const vector<Point3f> &oldData,
 		const vector<int> &planeIndices,
 		const int &minimumNumberOfPointsPerPlane,
 		map<int, int> &numberOfPointsPerPlane,
-		vector<Point3d> &newData,
+		vector<Point3f> &newData,
 		vector<int> &newPlaneIndices,
 		int &numberOfPlanes);
 
 /**
  * @brief This functions calculates the distances of each point to its respective plane
  e
- * @param [in] [vector<Point3d>] data - Set of all points
- * @param [in] [vector< vector<double> >] planeParameters - Set of plane parameters for the planes
+ * @param [in] [vector<Point3f>] data - Set of all points
+ * @param [in] [vector< vector<float> >] planeParameters - Set of plane parameters for the planes
  * 								planeParamters[i] corresponds to parameters of plane <i>i</i>
  * @param [in] [vector<int>] planeIndices - Indices revealing which point belongs
  * 										to which plane
- * @param [out] [vector<double>] distanceMatrix - Distance of each point from its respective plane
+ * @param [out] [vector<float>] distanceMatrix - Distance of each point from its respective plane
  * @param [out] [vector< vector<int> >] planePointsIndexMapping - This contains
  * 					which plane contains which points.
  * 					Example: planePointsIndexMapping[i] contains points corresponding to plane <i>i</i>
  * @return Nothing
  */
 void calculateDistanceFromPlane(
-		const vector<Point3d> &data,
-		const vector< vector<double> > &planeParameters,
+		const vector<Point3f> &data,
+		const vector< vector<float> > &planeParameters,
 		const vector<int> &planeIndices,
-		vector<double> &distanceMatrix,
+		vector<float> &distanceMatrix,
 		vector< vector<int> > &planePointsIndexMapping);
 
 /**
  *
- * @param [in] [vector<double>] distanceMatrix - Distance of each point from its respective plane
+ * @param [in] [vector<float>] distanceMatrix - Distance of each point from its respective plane
  * @param [in] [vector< vector<int> >] planePointsIndexMapping - This contains
  * 					which plane contains which points.
  * 					Example: planePointsIndexMapping[i] contains points corresponding to plane <i>i</i>
- * @param [out] [vector<Point3d>] newSortedData - Sorted planePoints according to planeIndex
- * @param [out] [vector< vector<double> >] newPlaneParameters - Plane parameters corresponding to index i.
+ * @param [out] [vector<Point3f>] newSortedData - Sorted planePoints according to planeIndex
+ * @param [out] [vector< vector<float> >] newPlaneParameters - Plane parameters corresponding to index i.
  * 					newPlaneParameters[i] contains parameters for plane i
  * @param [out] [map<int, pair<int, int> >] planeIndexBounds - The bounds which tells the plane(i) points are from index k1 to k2
  */
 void removePointsFarFromPlane(
-		const vector<Point3d> &data,
-		const vector< vector<double> > &planeParameters,
-		const vector<double> &distanceMatrix,
+		const vector<Point3f> &data,
+		const vector< vector<float> > &planeParameters,
+		const vector<float> &distanceMatrix,
 		const vector< vector<int> > &planePointsIndexMapping,
-		vector<Point3d> &newSortedData,
-		vector< vector<double> > &newPlaneParameters,
+		vector<Point3f> &newSortedData,
+		vector< vector<float> > &newPlaneParameters,
 		map<int, pair<int, int> > &planeIndexBounds);
 
 void removePointsFarFromPlane1(
-		const vector<Point3d> &data,
-		const vector< vector<double> > &planeParameters,
-		const vector<double> &distanceMatrix,
+		const vector<Point3f> &data,
+		const vector< vector<float> > &planeParameters,
+		const vector<float> &distanceMatrix,
 		const vector< vector<int> > &planePointsIndexMapping,
-		vector< vector<Point3d> > &newSortedData,
-		vector< vector<double> > &newPlaneParameters );
+		vector< vector<Point3f> > &newSortedData,
+		vector< vector<float> > &newPlaneParameters );
 
 void get3DPlaneProjectionsOfPoints (
-		const vector<Point3d> &data,
-		const vector< vector<double> > &planeParameters,
+		const vector<Point3f> &data,
+		const vector< vector<float> > &planeParameters,
 		const map<int, pair<int, int> > &planeIndexBounds,
-		vector<Point3d> &projectionsOf3DPoints	);
+		vector<Point3f> &projectionsOf3DPoints	);
 
 void get3DPlaneProjectionsOfPoints1 (
-		const vector< vector<Point3d> > &data,
-		const vector< vector<double> > &planeParameters,
-		vector<Point3d> &projectionsOf3DPoints	);
+		const vector< vector<Point3f> > &data,
+		const vector< vector<float> > &planeParameters,
+		vector<Point3f> &projectionsOf3DPoints	);
 
-double getKthPercentile(
-		const vector<double> &data,
+float getKthPercentile(
+		const vector<float> &data,
 		const int k);
 
 #endif /* ADDITIONALSTEPS_HPP_ */
